// src/redux/types.tsx
export enum ACTION_TYPES {
	GET_FETCH_DATA = 'APP/GET_FETCH_DATA',
	FETCH_DATA = 'APP/FETCH_DATA',
	REQUEST_DATA = 'APP/REQUEST_DATA',
	SHOW_LOADER = 'APP/SHOW_LOADER',
	HIDE_LOADER = 'APP/HIDE_LOADER',
}

export enum INNER_CONTENT_TYPES {
	REBUILD_DESIGN = 'APP/REBUILD_DESIGN',
	SHOW_BUTTON_LOADER = 'APP/BUTTON_LOADER',
	NOTIFICATION = 'APP/NOTIFICATION',
	ADD_PAGINATED_ITEMS = 'APP/ADD_PAGINATED_ITEMS',
}

export enum LOGIN_PAGE_TYPES {
	LOGIN_PAGE = 'APP/LOGIN_PAGE',
	USER_DATA = 'APP/USER_DATA',
	SHOW_ALERT = 'APP/SHOW_ALERT',
	HIDE_ALERT = 'APP/HIDE_ALERT',
}